fix(mod): provider 로직 이주
